% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadtest.R
\name{encode_html_entities}
\alias{encode_html_entities}
\title{Encode HTML/XML entities}
\usage{
encode_html_entities(stringified_body)
}
\arguments{
\item{stringified_body}{the request body as a single string}
}
\value{
a string
}
\description{
Jemeter's documentation in XML which means special characters like &, <, >
need to be encoded as HTML entities. This function does that without dependencies.
}
\examples{
encode_html_entities('{"title":"this & that"}')
}
