% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{java_path}
\alias{java_path}
\title{Command for the java executable}
\usage{
java_path()
}
\value{
A string with the command to run for jmeter
}
\description{
The command to call java. Usually just "java",
but you can use Sys.setenv("LOADTEST_JAVA_PATH") to specify the location if
java isn't in the path environment variable. This is only used for checking
that the package should work when loaded, so it's not important to change unless you hate the warning
any circumstance.
}
