% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{jmeter_path}
\alias{jmeter_path}
\title{Command for the jmeter executable}
\usage{
jmeter_path()
}
\value{
A string with the command to run for jmeter
}
\description{
The command to call to run jmeter. Usually just "jmeter",
but you can use Sys.setenv("LOADTEST_JMETER_PATH") to specify the location if
jmeter isn't in the path environment variable
}
