% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\docType{package}
\name{loadtest-package}
\alias{loadtest-package}
\title{loadtest: load testing directly from R}
\description{
This package has the loadtest() function, which allows you to run a load test for
an HTTP request, and returns a summary data frame. It also includes helper functions
to create plots and reports
}
