% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadtest_report.R
\name{loadtest_report}
\alias{loadtest_report}
\title{Create an output report of a jmeter run}
\usage{
loadtest_report(results, output_file)
}
\arguments{
\item{output_file}{the location to save the report}

\item{result}{the output of using loadtest()}
}
\description{
This function uses R markdown to take the results of a jmeter run and turn it
}
\examples{
results <- loadtest(url = "https://www.google.com", method="GET", threads = 3, loops = 5)
loadtest_report(results,"~/report.html")
}
