% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadtest.R
\name{parse_query_string}
\alias{parse_query_string}
\title{Convert a query string into its parts}
\usage{
parse_query_string(query_string)
}
\arguments{
\item{query_string}{a string containing the query parameters e.g.: "postId=1&userId=1"}
}
\value{
a list
}
\description{
This function gets the query parameters from a url as input,
and outputs a list. Each element's name is the name of the parameters,
and each element's value is the parameter's value.
}
\examples{
parse_query_string("postId=1&userId=1")
}
