% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadtest.R
\name{parse_url}
\alias{parse_url}
\title{Convert a url into core components}
\usage{
parse_url(url)
}
\arguments{
\item{url}{a string containing a url}
}
\description{
This code takes a url and breaks it into several components
\describe{
\item{protocol}{Either http or https}
\item{domain}{The base domain, such as https://www.t-mobile.com}
\item{path}{The path after the base domain such as /mail/account=1}
\item{port}{The port to use, either 80 for HTTP, 443 for HTTPS, or anything if explicitly set like :8000 after a domain}
}
}
