% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_elapsed_times}
\alias{plot_elapsed_times}
\title{Plot the elapsed times of the requests}
\usage{
plot_elapsed_times(results)
}
\arguments{
\item{results}{A data frame returned from the loadtest function}
}
\value{
A ggplot2 showing the elapsed times of the requests during the test
}
\description{
Plot the elapsed times of the requests
}
\examples{
results <- loadtest("google.com","GET")
plot_elapsed_times(result)
}
