% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_elapsed_times_histogram}
\alias{plot_elapsed_times_histogram}
\title{Plot the elapsed times of the requests as a histogram}
\usage{
plot_elapsed_times_histogram(results, binwidth = 250)
}
\arguments{
\item{results}{A data frame returned from the loadtest function}

\item{binwidth}{The binwidth for the histogram}
}
\value{
A ggplot2 showing the elapsed times of the requests during the test
}
\description{
Plot the elapsed times of the requests as a histogram
}
\examples{
results <- loadtest("google.com","GET")
plot_elapsed_times_histogram(results)
}
