% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_requests_by_thread}
\alias{plot_requests_by_thread}
\title{Plot the requests per second made during the test}
\usage{
plot_requests_by_thread(results)
}
\arguments{
\item{results}{A data frame returned from the loadtest function}
}
\value{
A ggplot2 showing how each thread's tests went
}
\description{
Plot the requests per second made during the test
}
\examples{
results <- loadtest("google.com","GET")
plot_requests_by_thread(results)
}
